+++ src/FunFair.Test.Infrastructure/Extensions/TestLoggerExtension.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 
@@ -7,39 +6,23 @@ namespace FunFair.Test.Infrastructure.Extensions;
 
 public static class TestLoggerExtension
 {
-    [SuppressMessage(
-        "Microsoft.Performance",
-        "CA1873: Evaluation of this argument may be expensive and unnecessary if logging is disabled",
-        Justification = "This is a unit test assembly - not so worried about performance"
-    )]
     public static void Received(this ILogger logger, LogLevel logLevel, string message, int received = 1)
     {
-        logger
-            .Received(received)
-            .Log(
-                logLevel: logLevel,
-                Arg.Any<EventId>(),
-                Arg.Is<object>(o => StringComparer.Ordinal.Equals(x: o.ToString(), y: message)),
-                Arg.Any<Exception?>(),
-                Arg.Any<Func<object, Exception?, string>>()
-            );
+        EventId eventId = Arg.Any<EventId>();
+        object state = Arg.Is<object>(o => StringComparer.Ordinal.Equals(x: o.ToString(), y: message));
+        Exception? exception = Arg.Any<Exception?>();
+        Func<object, Exception?, string> formatter = Arg.Any<Func<object, Exception?, string>>();
+
+        logger.Received(received).Log(logLevel: logLevel, eventId, state, exception, formatter);
     }
 
-    [SuppressMessage(
-        "Microsoft.Performance",
-        "CA1873: Evaluation of this argument may be expensive and unnecessary if logging is disabled",
-        Justification = "This is a unit test assembly - not so worried about performance"
-    )]
     public static void DidNotReceive(this ILogger logger, LogLevel logLevel, string message)
     {
-        logger
-            .DidNotReceive()
-            .Log(
-                logLevel: logLevel,
-                Arg.Any<EventId>(),
-                Arg.Is<object>(o => StringComparer.Ordinal.Equals(x: o.ToString(), y: message)),
-                Arg.Any<Exception?>(),
-                Arg.Any<Func<object, Exception?, string>>()
-            );
+        EventId eventId = Arg.Any<EventId>();
+        object state = Arg.Is<object>(o => StringComparer.Ordinal.Equals(x: o.ToString(), y: message));
+        Exception? exception = Arg.Any<Exception?>();
+        Func<object, Exception?, string> formatter = Arg.Any<Func<object, Exception?, string>>();
+
+        logger.DidNotReceive().Log(logLevel: logLevel, eventId, state, exception, formatter);
     }
 }